src/test/java/com/qa/ims/controllers/OrderControllerTest.java
package com.qa.ims.controllers;

import static org.junit.Assert.assertEquals;

import java.sql.Date;
import java.util.ArrayList;
import java.util.List;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.MockitoJUnitRunner;

import com.qa.ims.controller.OrderController;
import com.qa.ims.persistence.dao.ItemDAO;
import com.qa.ims.persistence.dao.OrderDAO;
import com.qa.ims.persistence.domain.Order;
import com.qa.ims.persistence.domain.OrderLine;
import com.qa.ims.utils.Utils;

@RunWith(MockitoJUnitRunner.class)
public class OrderControllerTest {

	@Mock
	private Utils utils;

	@Mock
	private OrderDAO dao;

	@Mock
	private ItemDAO iDao;

	@InjectMocks
	private OrderController controller;

	@Test
	public void testReadAll() {
		List<Order> orders = new ArrayList<>();
		Date date = Date.valueOf("1996-04-04");
		List<OrderLine> orderlines = new ArrayList<>();
		orderlines.add(new OrderLine(1l, 1l, 1l));
		Order order = new Order(1L, 1l, date, 1.00, orderlines);
		orders.add(order);

		Mockito.when(dao.readAll()).thenReturn(orders);

		assertEquals(orders, controller.readAll());

		Mockito.verify(dao, Mockito.times(1)).readAll();

	}
	
	//  tostring is null causing this to fail  
	//  @Test public void createTest() { 
	//	  Date date = Date.valueOf("1996-04-04"); 
	//	  final long cid = 1l;
	//	  final Order created = new Order(cid, date);
	//	  final Order expected = new Order(1l, date);
	  
	//	  Mockito.when(utils.getLong()).thenReturn(cid);
	//	  Mockito.when(dao.create(expected)).thenReturn(created);
	  
	//	  assertEquals(created, controller.create());
		  
	//	  Mockito.verify(utils, Mockito.times(1)).getLong();
	//	  Mockito.verify(dao, Mockito.times(1)).create(created);
	  
	//  }
	 
		/*
		 * @Test public void updateTest() { Long id =1l; Double total = 25.00; String
		 * date = "1996-04-04";
		 * 
		 * Order toUpdate = new Order(2L, id, null, total, date); Order updated = new
		 * Order(2l, 1L, null, 25.00, "1996-04-04");
		 * 
		 * Mockito.when(utils.getLong()).thenReturn(id);
		 * Mockito.when(utils.getDouble()).thenReturn(total);
		 * Mockito.when(utils.getString()).thenReturn(date);
		 * Mockito.when(dao.create(toUpdate)).thenReturn(updated);
		 * 
		 * assertEquals(updated, controller.update()); }
		 */
	 

	@Test
	public void testDelete() {
		final long ID = 1L;

		Mockito.when(utils.getLong()).thenReturn(ID);
		Mockito.when(dao.delete(ID)).thenReturn(1);

		assertEquals(1L, this.controller.delete());

		Mockito.verify(utils, Mockito.times(1)).getLong();
		Mockito.verify(dao, Mockito.times(1)).delete(ID);
	}

}
